+++ client/src/app/app/_editor.tsx
@@ -1,4 +1,6 @@
+import { useEffect } from "react"
 import { createFileRoute, Outlet, redirect } from "@tanstack/react-router"
+import { useAtom } from "jotai"
 
 import { EditorHeaderAndSidebar } from "@/components/editor/Controls"
 import { Sidebar } from "@/components/editor/EditorSidebar"
@@ -6,6 +8,8 @@ import { StatusBar } from "@/components/editor/StatusBar"
 import { ScrollArea } from "@/components/ui/scroll-area"
 import { TooltipProvider } from "@/components/ui/tooltip"
 import { isAuthenticated } from "@/lib/auth"
+import { useSupa } from "@/lib/supabaseClient"
+import { userSettingsStore } from "@/store/userSettingsStore"
 
 export const Route = createFileRoute("/app/_editor")({
   component: BaseLayout,
@@ -22,6 +26,19 @@ export const Route = createFileRoute("/app/_editor")({
 })
 
 function BaseLayout() {
+  const [userSettings, setUserSettings] = useAtom(userSettingsStore)
+  const supa = useSupa()
+
+  const setUser = async () => {
+    const {
+      data: { user },
+    } = await supa.auth.getUser()
+    setUserSettings({ ...userSettings, userId: user?.id || null })
+  }
+
+  useEffect(() => {
+    setUser()
+  }, [])
   return (
     <TooltipProvider>
       <div className="flex h-dvh w-full flex-col overflow-hidden">
@@ -34,7 +51,7 @@ function BaseLayout() {
           <main className="flex flex-1 overflow-x-hidden">
             <div className="flex min-h-screen w-full flex-col bg-muted/40">
               <Sidebar />
-              <div className="ml-12">
+              <div className="sm:ml-12">
                 <Outlet />
               </div>
             </div>
