+++ client/src/app/app/_editor/index.tsx
@@ -1,9 +1,9 @@
 import { createFileRoute } from "@tanstack/react-router"
 
-import { Editor } from "@/components/editor/Canvas"
+import { Canvas } from "@/components/editor/Canvas"
 
 // import { Seo } from "@/components/layout/seo"
 
 export const Route = createFileRoute("/app/_editor/")({
-  component: Editor,
+  component: Canvas,
 })
